Commitando finalização da API, incluindo cadastros de pautas, associados, sessão, votos, e teste do voto

+++ src/main/java/br/com/votacao/votaweb/controller/AssociadoController.java
@@ -35,7 +35,7 @@ public class AssociadoController {
 	}
 
     @GetMapping("/v1/associados/{id}")
-    public Optional<Associado> buscaPorId(@PathVariable Long id) {
+    public Optional<Associado> buscaPorId(@PathVariable int id) {
         return associadoService.findById(id);
     }
